+++ spec/integrations/fixtures/rspec/let_it_be_modification_detection_fixture.rb
@@ -5,6 +5,12 @@ require_relative "../../../support/transactional_context"
 
 require "test_prof/recipes/rspec/let_it_be"
 
+RSpec.configure do |config|
+  config.define_derived_metadata(let_it_be_frost: true) do |metadata|
+    metadata[:let_it_be_modifiers] ||= {freeze: true}
+  end
+end
+
 RSpec.describe "Modification detection", :transactional, let_it_be_frost: true do
   include TestProf::FactoryBot::Syntax::Methods
 
@@ -18,7 +24,7 @@ RSpec.describe "Modification detection", :transactional, let_it_be_frost: true d
 
     it "detects the leak" do
       expect { user.update!(name: "John Doe") }
-        .to raise_error(/[Cc]an't modify frozen/)
+        .to raise_error(/can't modify frozen/i)
     end
 
     it { expect(user.name).to eq("Original Name") }
@@ -29,14 +35,14 @@ RSpec.describe "Modification detection", :transactional, let_it_be_frost: true d
 
     it "detects the leak" do
       expect { post.user.update!(name: "John Doe") }
-        .to raise_error(/[Cc]an't modify frozen/)
+        .to raise_error(/can't modify frozen/i)
     end
 
     it { expect(post.user.name).to eq("Original Name") }
   end
 
-  describe "no state leakage with transactional tests with `refind: true`", order: :defined do
-    let_it_be(:post, refind: true) { create(:post, user: create(:user, name: "Original Name")) }
+  describe "no state leakage with transactional tests with `refind: true` and `freeze: false`", order: :defined do
+    let_it_be(:post, refind: true, freeze: false) { create(:post, user: create(:user, name: "Original Name")) }
 
     it "leaks" do
       post.user.update!(name: "John Doe")
@@ -45,8 +51,8 @@ RSpec.describe "Modification detection", :transactional, let_it_be_frost: true d
     it { expect(post.user.name).to eq("Original Name") }
   end
 
-  describe "no state leakage with transactional tests with `reload: true`", order: :defined do
-    let_it_be(:post, reload: true) { create(:post, user: create(:user, name: "Original Name")) }
+  describe "no state leakage with transactional tests with `reload: true` and `freeze: false`", order: :defined do
+    let_it_be(:post, reload: true, freeze: false) { create(:post, user: create(:user, name: "Original Name")) }
 
     it "leaks" do
       post.user.update!(name: "John Doe")
@@ -60,47 +66,25 @@ RSpec.describe "Modification detection", :transactional, let_it_be_frost: true d
 
     it "detects the leak in an array item" do
       expect { users.first.update!(name: "John Doe") }
-        .to raise_error(/[Cc]an't modify frozen/)
+        .to raise_error(/can't modify frozen/i)
     end
 
     it "detects the leak in the array itself" do
       expect { users << "yet another user" }
-        .to raise_error(/[Cc]an't modify frozen/)
+        .to raise_error(/can't modify frozen/i)
     end
   end
 
   describe "infers `freeze: false`" do
-    context "from `reload: true`" do
-      let_it_be(:user, reload: true) { create(:user) }
-      let_it_be(:users) { [user] }
+    let_it_be(:user, freeze: false) { create(:user) }
+    let_it_be(:users) { [user] }
 
-      it "skips leakage detection" do
-        expect { user.update!(name: "Other Name") }
-          .not_to raise_error
-      end
+    it "skips leakage detection" do
+      expect { user.update!(name: "Other Name") }
+        .not_to raise_error
     end
 
-    context "from `refind: true`" do
-      let_it_be(:user, refind: true) { create(:user) }
-      let_it_be(:users) { [user] }
-
-      it "skips leakage detection" do
-        expect { user.update!(name: "Other Name") }
-          .not_to raise_error
-      end
-    end
-
-    context "from `freeze: false`" do
-      let_it_be(:user, freeze: false) { create(:user) }
-      let_it_be(:users) { [user] }
-
-      it "skips leakage detection" do
-        expect { user.update!(name: "Other Name") }
-          .not_to raise_error
-      end
-    end
-
-    context "from metadata", let_it_be_frost: false do
+    context "from metadata", let_it_be_modifiers: {freeze: false} do
       let_it_be(:user) { create(:user) }
       let_it_be(:users) { [user] }
 
@@ -123,9 +107,9 @@ RSpec.describe "Modification detection", :transactional, let_it_be_frost: true d
 
           it "only freezes what's necessary" do
             expect { one.push(1) }.not_to raise_error
-            expect { two.push(2) }.to raise_error(/[Cc]an't modify frozen/)
+            expect { two.push(2) }.to raise_error(/can't modify frozen/i)
             expect { two.first.push(2) }.not_to raise_error
-            expect { two.last.push(2) }.to raise_error(/[Cc]an't modify frozen/)
+            expect { two.last.push(2) }.to raise_error(/can't modify frozen/i)
             expect { three.push(3) }.not_to raise_error
           end
         end